# FIX-#0000: optimize `__getitem__` flow of loc/iloc